+++ pyrk/utilities/exceptions.py
